personal_knowledge_rag_demo_fixed/requirements.txt
streamlit==1.50.0
chromadb==1.5.7
sentence-transformers==5.1.2
pandas==2.3.3
numpy<2
urllib3<2
python-dotenv==1.2.1
openai==2.31.0
ragas>=0.4,<0.5
rapidfuzz>=3.0.0
